requirements.txt
nltk
numpy
tensorflow
keras
pillow
tk